Moved to RV64I, executor fixes, memory via mmap, 8 queens added
TODO: more accurate trasition to RV64I with naming, readme update, binary instructions blocks cache, not pointers

+++ src/hart/hart.hpp
@@ -8,7 +8,7 @@
 #include "decode_execute_module/common.hpp"
 #include "block_cache.hpp"
 
-using reg_t = uint32_t;
+using reg_t = uint64_t;
 
 class Hart {
 public:
